+++ app/components/Header/header.jsx
@@ -1,0 +1,34 @@
+import Link from 'next/link'
+import styles from './Header.module.css'
+
+const links = [{
+  label: 'Home',
+  route: '/'
+},
+{
+  label: 'About',
+  route: '/about'
+},
+{
+  label: 'Posts',
+  route: '/posts'
+}
+]
+function Header () {
+  return (
+    <header className={styles.header}>
+      <nav>
+        <ul className={styles.navigation}>
+          {links.map(({ label, route }) => (
+            <li className={styles.links} key={route}>
+              <Link href={route}>
+                {label}
+              </Link>
+            </li>
+          ))}
+        </ul>
+      </nav>
+    </header>
+  )
+}
+export default Header
