--- app/components/Navigation/Navigation.module.css
@@ -1,21 +1,0 @@
-.header {
-    border: 1px solid #f0f0ff;
-    border-radius: 4px;
-    padding: 4px;
-}
-
-.navigation {
-     display: flex;
-     list-style: none;
-     gap: 10px;
-     margin: 5px;
-}
-.navigation a:hover {
-    color: #f0f0ff;
-    text-decoration: underline;
-}
-
-.links {
-    display: flex;
-    flex-direction: row;
-}
